[Mezuro] renaming kalibro configurarion variable names to avoid conflits with solr configuration variable names

+++ plugins/mezuro/lib/mezuro_plugin/configuration_content.rb
@@ -3,8 +3,8 @@ class MezuroPlugin::ConfigurationContent < Article
 
   settings_items :description, :configuration_to_clone_name
 
-  after_save :send_configuration_to_service
-  after_destroy :remove_configuration_from_service
+  after_save :send_kalibro_configuration_to_service
+  after_destroy :remove_kalibro_configuration_from_service
 
   def self.short_description
     'Kalibro configuration'
@@ -21,54 +21,55 @@ class MezuroPlugin::ConfigurationContent < Article
     end
   end
 
-  def configuration
-    @configuration ||= Kalibro::Configuration.find_by_name(self.name)
-    if @configuration.nil? 
-      errors.add_to_base("Kalibro Configuration not found")
+  def kalibro_configuration
+    begin
+      @kalibro_configuration ||= Kalibro::Configuration.find_by_name(self.name)
+    rescue Exception => exception 
+      errors.add_to_base(exception.message)
     end
-    @configuration
+    @kalibro_configuration
   end
 
   def metric_configurations
-    configuration.metric_configurations
+    kalibro_configuration.metric_configurations
   end
 
-  def configuration_names
+  def kalibro_configuration_names
     ["None"] + Kalibro::Configuration.all_names.sort
   end
 
   private
 
   def validate_kalibro_configuration_name
-    existing = configuration_names.map { |a| a.downcase}
+    existing = kalibro_configuration_names.map { |a| a.downcase}
 
     if existing.include?(name.downcase)
       errors.add_to_base("Configuration name already exists in Kalibro")
     end
   end
 
-  def send_configuration_to_service
-    if editing_configuration?
-      configuration.update_attributes({:description => description})
+  def send_kalibro_configuration_to_service
+    if editing_kalibro_configuration?
+      kalibro_configuration.update_attributes({:description => description})
     else
       create_kalibro_configuration
     end
   end
 
-  def remove_configuration_from_service
-    configuration.destroy
+  def remove_kalibro_configuration_from_service
+    kalibro_configuration.destroy unless kalibro_configuration.nil?
   end
 
   def create_kalibro_configuration
     attributes = {:name => name, :description => description}
-    if cloning_configuration?
+    if cloning_kalibro_configuration?
       attributes[:metric_configuration] = configuration_to_clone.metric_configurations_hash
     end
     Kalibro::Configuration.create attributes
   end
   
-  def editing_configuration?
-    configuration.present?
+  def editing_kalibro_configuration?
+    kalibro_configuration.present?
   end
   
   def configuration_to_clone
@@ -76,10 +77,10 @@ class MezuroPlugin::ConfigurationContent < Article
   end
   
   def find_configuration_to_clone
-    configuration_to_clone_name.nil? ? nil : Kalibro::Configuration.find_by_name(configuration_to_clone_name)
+    (configuration_to_clone_name == "None") ? nil : Kalibro::Configuration.find_by_name(configuration_to_clone_name)
   end
   
-  def cloning_configuration?
+  def cloning_kalibro_configuration?
     configuration_to_clone.present?
   end
